Refactor: move persona slash commands
